Update compilation environment
