+++ Tests/Fixtures/CompilationEnvironment/CoreData-Definitions.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum IntegerEnum: Int16 {
+  case test1
+}
+public enum StringEnum: String {
+  case test1
+}
