+++ Tests/Fixtures/CompilationEnvironment/CoreData-ModuleTypes.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+public final class CustomArray: NSArray {}
+public final class CustomPolyline: NSObject {}
