+++ Tests/Fixtures/CompilationEnvironment/IB-iOS-Definitions.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+public final class CreateAccViewController: UIViewController {}
+public final class PickerViewController: UIViewController {}
+public final class SlideLeftSegue: UIStoryboardSegue {}
+public final class SlideUpSegue: UIStoryboardSegue {}
