+++ Tests/Fixtures/CompilationEnvironment/IB-iOS-Login.swift
@@ -1,0 +1,4 @@
+import UIKit
+
+public final class SlideDownSegue: UIStoryboardSegue {}
+public final class ValidatePasswordViewController: UIViewController {}
