+++ Tests/Fixtures/CompilationEnvironment/IB-iOS-NamespacedLogin.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+enum ExtraModule {
+  final class SlideDownSegue: UIStoryboardSegue {}
+  final class ValidatePasswordViewController: UIViewController {}
+}
