+++ Tests/Fixtures/CompilationEnvironment/IB-macOS-Definitions.swift
@@ -1,0 +1,6 @@
+import Cocoa
+
+public final class CustomTabViewController: NSWindowController {}
+public final class DetailsViewController: NSWindowController {}
+public final class RotateSegue: NSStoryboardSegue {}
+public final class ZoomSegue: NSStoryboardSegue {}
