+++ Tests/Fixtures/CompilationEnvironment/IB-macOS-Login.swift
@@ -1,0 +1,4 @@
+import Cocoa
+
+public final class LoginSegue: NSStoryboardSegue {}
+public final class LoginViewController: NSWindowController {}
